+++ src/onegov.feriennet/onegov/feriennet/tests/test_views.py
@@ -3,6 +3,7 @@ import onegov.feriennet
 import transaction
 
 from datetime import datetime, timedelta, date
+from freezegun import freeze_time
 from onegov.activity import ActivityCollection
 from onegov.activity import AttendeeCollection
 from onegov.activity import BookingCollection
@@ -915,6 +916,59 @@ def test_enroll_age_mismatch(feriennet_app):
 
     page.form['ignore_age'] = True
     assert "Wunschliste hinzugefügt" in page.form.submit().follow()
+
+
+def test_enroll_after_wishlist_phase(feriennet_app):
+    activities = ActivityCollection(feriennet_app.session(), type='vacation')
+    periods = PeriodCollection(feriennet_app.session())
+    occasions = OccasionCollection(feriennet_app.session())
+
+    retreat = activities.add("Retreat", username='admin@example.org')
+    retreat.propose().accept()
+
+    prebooking = tuple(d.date() for d in (
+        datetime.now() - timedelta(days=1),
+        datetime.now() + timedelta(days=1)
+    ))
+
+    execution = tuple(d.date() for d in (
+        datetime.now() + timedelta(days=10),
+        datetime.now() + timedelta(days=12)
+    ))
+
+    occasions.add(
+        start=datetime(2016, 10, 8, 8),
+        end=datetime(2016, 10, 9, 16),
+        age=(5, 10),
+        timezone="Europe/Zurich",
+        activity=retreat,
+        period=periods.add(
+            title="2016",
+            prebooking=prebooking,
+            execution=execution,
+            active=True
+        )
+    )
+
+    UserCollection(feriennet_app.session()).add(
+        'member@example.org', 'hunter2', 'member')
+
+    transaction.commit()
+
+    admin = Client(feriennet_app)
+    admin.login_admin()
+
+    fill_out_profile(admin)
+
+    page = admin.get('/angebot/retreat').click("Anmelden")
+    page.form['first_name'] = "Tom"
+    page.form['last_name'] = "Sawyer"
+    page.form['gender'] = 'male'
+    page.form['birth_date'] = "2015-01-01"
+    page.form['ignore_age'] = True
+
+    with freeze_time(datetime.now() + timedelta(days=2)):
+        assert "nur während der Wunschphase" in page.form.submit()
 
 
 def test_booking_view(feriennet_app):